feat(ToggleSwitch): add support for a `ref` on the inner <button> element
